Code splitting for Total Values and Profit Gross calculations. Some minor data output and formating fixes
