+++ package.json
@@ -4,7 +4,8 @@
   "description": "RewardPay coding challenge by Kateryna Bilokur",
   "main": "index.js",
   "scripts": {
-    "test": "vitest",
+    "test-watch": "vitest",
+    "test": "vitest run",
     "dev": "nodemon src/index.js",
     "start": "node src/index.js"
   },
