+++ src/calculations/profitMetrics.js
@@ -1,28 +1,4 @@
-function totalValueByCategory(category, data) {
-  return data
-    .filter(
-      (item) => item.account_category?.toLowerCase() === category.toLowerCase()
-    )
-    .reduce((sum, item) => sum + (+item.total_value || 0), 0);
-}
-
-/**
- * Calculates the total Revenue value.
- * @param {Array} data - The array of data to calculate from.
- * @returns {number} - The total Revenue.
- */
-function calculateRevenue(data) {
-  return totalValueByCategory("revenue", data);
-}
-
-/**
- * Calculates the total Expense value.
- * @param {Array} data - The array of data to calculate from.
- * @returns {number} - The total Expense.
- */
-function calculateExpenses(data) {
-  return totalValueByCategory("expense", data);
-}
+const { calculateRevenue, calculateExpenses } = require("./totalValueMetrics");
 
 /**
  * Calculates the total Gross Profit Margin value.
@@ -59,8 +35,6 @@ function calculateNetProfitMargin(data, revenue = null, expenses = null) {
 }
 
 module.exports = {
-  calculateRevenue,
-  calculateExpenses,
   calculateGrossProfitMargin,
   calculateNetProfitMargin,
 };
